Initial button implementation

+++ src/ui/rectangle/Rectangle.cpp
@@ -2,7 +2,7 @@
 
 namespace Chess::Rendering::Elements {
 
-bool Rectangle::IsWithinRect(int x, int y) {
+bool Rectangle::IsWithinRect(float x, float y) {
   // #region IsWithinRect
   if (x < rect.x) return false;
   if (x > rect.x + rect.w) return false;
@@ -12,7 +12,7 @@ bool Rectangle::IsWithinRect(int x, int y) {
   // #endregion
 }
 
-void Rectangle::SetPosition(int x, int y) {
+void Rectangle::SetPosition(float x, float y) {
   // #region SetPosition
   rect.x = x;
   rect.y = y;
